requirements.txt
alembic==1.18.4
bcrypt==4.2.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
gunicorn==22.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.0.3
argon2-cffi==23.1.0
argon2-cffi-bindings==23.1.0